Refactoring petclinic structure to allow horizontal slice by domain
