+++ petclinic-web/pom.xml
@@ -7,18 +7,29 @@
 		<artifactId>nimbus-parent</artifactId>
 		<version>1.2.0.BUILD-SNAPSHOT</version>
 	</parent>
+	<!-- <parent>
+		<groupId>org.springframework.boot</groupId>
+		<artifactId>spring-boot-starter-parent</artifactId>
+		<version>2.0.6.RELEASE</version>
+		<relativePath/> lookup parent from repository
+	</parent> -->
   <properties>
   	<nimbus.version>1.2.0.BUILD-SNAPSHOT</nimbus.version>
+  	<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
+		<java.version>1.8</java.version>
+		<spring-cloud.version>Finchley.SR2</spring-cloud.version>
   </properties>
   <groupId>com.anthem.petclinic.web</groupId>
   <artifactId>petclinic-web</artifactId>
   <packaging>war</packaging>
-  <version>1.0.0-SNAPSHOT</version>
+  <version>1.2.0-SNAPSHOT</version>
   <description>Petclinic Web Application</description>
   <dependencies>
   	<dependency>
 		<groupId>com.antheminc.oss</groupId>
 		<artifactId>nimbus-core</artifactId>
+		<version>${nimbus.version}</version>
 		<exclusions>
 			<exclusion>
 				 <groupId>de.flapdoodle.embed</groupId>
@@ -29,16 +40,18 @@
 	<dependency>
 		<groupId>com.antheminc.oss</groupId>
 		<artifactId>nimbus-ui</artifactId>
+		<version>${nimbus.version}</version>
 	</dependency>
 	<dependency>
 		<groupId>com.antheminc.oss</groupId>
 		<artifactId>nimbus-test</artifactId>
+		<version>${nimbus.version}</version>
 		<scope>test</scope>
 	</dependency>
 	<dependency>
 		<groupId>com.anthem.client.extension</groupId>
 		<artifactId>com.anthem.client.extension.petclinic</artifactId>
-		<version>1.0.0-SNAPSHOT</version>
+		<version>1.2.0-SNAPSHOT</version>
 		<exclusions>
 			<exclusion>
 				 <groupId>de.flapdoodle.embed</groupId>
@@ -60,6 +73,23 @@
 		<artifactId>spring-boot-configuration-processor</artifactId>
 	</dependency>
 	<dependency>
+			<groupId>org.springframework.cloud</groupId>
+			<artifactId>spring-cloud-starter-netflix-eureka-client</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>org.springframework.cloud</groupId>
+			<artifactId>spring-cloud-starter-netflix-ribbon</artifactId>
+		</dependency>
+	<dependency>
+		<groupId>org.springframework.boot</groupId>
+		<artifactId>spring-boot-starter-webflux</artifactId>
+	</dependency>
+		<dependency>
+			<groupId>io.projectreactor</groupId>
+			<artifactId>reactor-test</artifactId>
+			<scope>test</scope>
+		</dependency>
+	<dependency> 
 		<groupId>junit</groupId>
 		<artifactId>junit</artifactId>
 	</dependency>
@@ -123,8 +153,20 @@
 	<dependency>
 	 	<groupId>io.jsonwebtoken</groupId>
 		<artifactId>jjwt</artifactId>
+		<version>0.6.0</version>
 	</dependency>	
   </dependencies>
+  <!-- <dependencyManagement>
+		<dependencies>
+			<dependency>
+				<groupId>org.springframework.cloud</groupId>
+				<artifactId>spring-cloud-dependencies</artifactId>
+				<version>${spring-cloud.version}</version>
+				<type>pom</type>
+				<scope>import</scope>
+			</dependency>
+		</dependencies>
+	</dependencyManagement> -->
   <build>
     <resources>
        <resource>
