+++ platform.client.extension.petclinic/src/main/java/com/atlas/client/extension/petclinic/core/owner/OwnerCall.java
@@ -13,7 +13,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package com.atlas.client.extension.petclinic.core;
+package com.atlas.client.extension.petclinic.core.owner;
 
 import java.time.LocalDateTime;
