+++ platform.client.extension.petclinic/src/main/java/com/atlas/client/extension/petclinic/core/owner/Visit.java
@@ -1,4 +1,4 @@
-package com.atlas.client.extension.petclinic.core;
+package com.atlas.client.extension.petclinic.core.owner;
 
 import java.time.LocalDate;
