+++ platform.client.extension.petclinic/src/main/java/com/atlas/client/extension/petclinic/core/pet/Pet.java
@@ -1,4 +1,4 @@
-package com.atlas.client.extension.petclinic.core;
+package com.atlas.client.extension.petclinic.core.pet;
 
 import java.time.LocalDate;
 import java.util.List;
