+++ platform.client.extension.petclinic/src/main/java/com/atlas/client/extension/petclinic/core/veterinarian/Veterinarian.java
@@ -1,4 +1,4 @@
-package com.atlas.client.extension.petclinic.core;
+package com.atlas.client.extension.petclinic.core.veterinarian;
 
 import java.util.List;
 import java.util.Optional;
