+++ platform.client.extension.petclinic/src/main/java/com/atlas/client/extension/petclinic/view/home/NoteLineItem.java
@@ -23,7 +23,7 @@ import com.antheminc.oss.nimbus.domain.defn.ViewConfig.GridColumn;
 import com.antheminc.oss.nimbus.domain.defn.ViewConfig.Link;
 import com.antheminc.oss.nimbus.domain.defn.ViewConfig.LinkMenu;
 import com.antheminc.oss.nimbus.domain.defn.extension.Content.Label;
-import com.atlas.client.extension.petclinic.core.Note;
+import com.atlas.client.extension.petclinic.core.home.Note;
 
 import lombok.Getter;
 import lombok.Setter;
