+++ platform.client.extension.petclinic/src/main/java/com/atlas/client/extension/petclinic/view/home/VRDashboard.java
@@ -1,4 +1,4 @@
-package com.atlas.client.extension.petclinic.view;
+package com.atlas.client.extension.petclinic.view.home;
 
 import java.util.List;
 
@@ -16,7 +16,6 @@ import com.antheminc.oss.nimbus.domain.defn.ViewConfig.Section;
 import com.antheminc.oss.nimbus.domain.defn.ViewConfig.Tile;
 import com.antheminc.oss.nimbus.domain.defn.ViewConfig.ViewRoot;
 import com.antheminc.oss.nimbus.domain.defn.extension.Content.Label;
-import com.atlas.client.extension.petclinic.view.home.VPNotes;
 import com.atlas.client.extension.petclinic.view.owner.VisitLineItem;
 
 import lombok.Getter;
