+++ platform.client.extension.petclinic/src/main/java/com/atlas/client/extension/petclinic/view/owner/PetLineItemOwnerLanding.java
@@ -4,7 +4,7 @@ import com.antheminc.oss.nimbus.domain.defn.MapsTo;
 import com.antheminc.oss.nimbus.domain.defn.MapsTo.Path;
 import com.antheminc.oss.nimbus.domain.defn.ViewConfig.GridColumn;
 import com.antheminc.oss.nimbus.domain.defn.extension.Content.Label;
-import com.atlas.client.extension.petclinic.core.Pet;
+import com.atlas.client.extension.petclinic.core.pet.Pet;
 
 import lombok.Getter;
 import lombok.Setter;
