+++ platform.client.extension.petclinic/src/main/java/com/atlas/client/extension/petclinic/view/owner/VRVisitLanding.java
@@ -6,7 +6,7 @@ import com.antheminc.oss.nimbus.domain.defn.MapsTo;
 import com.antheminc.oss.nimbus.domain.defn.Repo;
 import com.antheminc.oss.nimbus.domain.defn.ViewConfig.Page;
 import com.antheminc.oss.nimbus.domain.defn.ViewConfig.ViewRoot;
-import com.atlas.client.extension.petclinic.core.Visit;
+import com.atlas.client.extension.petclinic.core.owner.Visit;
 
 import lombok.Getter;
 import lombok.Setter;
