+++ platform.client.extension.petclinic/src/main/java/com/atlas/client/extension/petclinic/view/owner/VisitLineItem.java
@@ -10,7 +10,7 @@ import com.antheminc.oss.nimbus.domain.defn.ViewConfig.GridColumn;
 import com.antheminc.oss.nimbus.domain.defn.ViewConfig.Link;
 import com.antheminc.oss.nimbus.domain.defn.ViewConfig.LinkMenu;
 import com.antheminc.oss.nimbus.domain.defn.extension.Content.Label;
-import com.atlas.client.extension.petclinic.core.Visit;
+import com.atlas.client.extension.petclinic.core.owner.Visit;
 
 import lombok.Getter;
 import lombok.Setter;
