+++ platform.client.extension.petclinic/src/main/java/com/atlas/client/extension/petclinic/view/pet/MealInstructionLineItem.java
@@ -31,7 +31,7 @@ import com.antheminc.oss.nimbus.domain.defn.ViewConfig.GridRowBody;
 import com.antheminc.oss.nimbus.domain.defn.ViewConfig.Link;
 import com.antheminc.oss.nimbus.domain.defn.ViewConfig.LinkMenu;
 import com.antheminc.oss.nimbus.domain.defn.extension.Content.Label;
-import com.atlas.client.extension.petclinic.core.MealInstruction;
+import com.atlas.client.extension.petclinic.core.pet.MealInstruction;
 
 import lombok.Getter;
 import lombok.Setter;
