+++ platform.client.extension.petclinic/src/main/java/com/atlas/client/extension/petclinic/view/pet/VPPetInfo.java
@@ -16,7 +16,7 @@ import com.antheminc.oss.nimbus.domain.defn.ViewConfig.Grid;
 import com.antheminc.oss.nimbus.domain.defn.ViewConfig.Section;
 import com.antheminc.oss.nimbus.domain.defn.ViewConfig.Tile;
 import com.antheminc.oss.nimbus.domain.defn.extension.Content.Label;
-import com.atlas.client.extension.petclinic.core.Pet;
+import com.atlas.client.extension.petclinic.core.pet.Pet;
 import com.atlas.client.extension.petclinic.view.owner.VisitLineItem;
 
 import lombok.Getter;
