+++ platform.client.extension.petclinic/src/main/java/com/atlas/client/extension/petclinic/view/veterinarian/VRVeterinarianLanding.java
@@ -15,7 +15,7 @@ import com.antheminc.oss.nimbus.domain.defn.ViewConfig.Section;
 import com.antheminc.oss.nimbus.domain.defn.ViewConfig.Tile;
 import com.antheminc.oss.nimbus.domain.defn.ViewConfig.ViewRoot;
 import com.antheminc.oss.nimbus.domain.defn.extension.Content.Label;
-import com.atlas.client.extension.petclinic.core.Veterinarian;
+import com.atlas.client.extension.petclinic.core.veterinarian.Veterinarian;
 
 import lombok.Getter;
 import lombok.Setter;
