+++ platform.client.extension.petclinic/src/test/java/com/atlas/client/extension/petclinic/scenariotests/pet/AddEditPetUnitTest.java
@@ -19,9 +19,9 @@ import com.antheminc.oss.nimbus.domain.model.state.EntityState.Param;
 import com.antheminc.oss.nimbus.support.Holder;
 import com.antheminc.oss.nimbus.test.domain.support.utils.ExtractResponseOutputUtils;
 import com.antheminc.oss.nimbus.test.domain.support.utils.MockHttpRequestBuilder;
-import com.atlas.client.extension.petclinic.core.Owner;
-import com.atlas.client.extension.petclinic.core.Pet;
-import com.atlas.client.extension.petclinic.core.Veterinarian;
+import com.atlas.client.extension.petclinic.core.owner.Owner;
+import com.atlas.client.extension.petclinic.core.pet.Pet;
+import com.atlas.client.extension.petclinic.core.veterinarian.Veterinarian;
 import com.atlas.client.extension.petclinic.scenariotests.AbstractPetclinicSpringTest;
 import com.atlas.client.extension.petclinic.view.pet.VPAddEditPet.VFAddEditPet;
 import com.atlas.client.extension.petclinic.view.pet.VRPet;
